adds favicon and site title
