+++ src/Components/MobileNav/MobileNav.module.scss
@@ -5,9 +5,13 @@
   height: 92vh;
   width: 100%;
   position: fixed;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
   top: 0;
   left: 0;
-  background: red;
+  background: $off-white;
   z-index: 999;
 }
