+++ src/Components/Navigation/Navigation.jsx
@@ -8,7 +8,6 @@ export default function Navigation({ isAuth }) {
       <Link to="/">home</Link>
       {isAuth && <Link to="/editor">editor</Link>}
       {isAuth && <Link to="/dashboard">dashboard</Link>}
-      <Link to="/design-system">design system</Link>
       <Link to="/templates">templates</Link>
     </nav>
   );
